pyproject.toml
[project]
name = "phone-compliance-system"
version = "0.1.0"
description = "Data Engineering Challenge: Client Phone Consent System"
requires-python = ">=3.9"
dependencies = [
    "pyspark>=3.5,<4",
]

[project.optional-dependencies]
dev = ["pytest>=8"]
cloud = ["delta-spark>=3.1"]

[tool.setuptools.packages.find]
include = ["pipeline*", "tests*"]

[tool.pytest.ini_options]
markers = [
    "integration: Cross-layer contract and consistency tests",
    "e2e: End-to-end tests against generated or sample data (opt-in)",
]
addopts = "-m 'not e2e'"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
